Reorganize acceptance tests and add APM tests
